server/requirements.txt
fastapi
python-dotenv==0.19.0
stripe==2.60.0
